requirements.txt
gitdb==4.0.2
GitPython==3.1.0
smmap==3.0.1
